+++ dangerfile.js
@@ -3,7 +3,11 @@ import { danger, fail } from 'danger';
 
 const lernaConfig = require('./lerna.json');
 
-const ignorePatterns = lernaConfig.command.publish.ignoreChanges || [];
+let ignorePatterns = [];
+
+if (lernaConfig.command) {
+  ignorePatterns = lernaConfig.command.publish.ignoreChanges;
+}
 
 const CHANGELOG_PATTERN = /^packages\/([a-z-])*\/CHANGELOG\.md/i;