+++ packages/webpack-config-terra/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
-* Changed
-  * Upgraded from `node-sass` to `sass (dart-sass)`
+## 2.0.0 - (August 13, 2021)
+
+* Breaking
+  * Upgraded from `node-sass` to `sass (dart-sass)`.
+  * Removed polyfill entries.
 
 ## 1.3.1 - (January 27, 2021)
 
@@ -14,7 +17,7 @@
 
 * Added
   * Added `browser` to `resolve.mainFields` in webpack module list.
-  
+
 * Changed
   * Passively consume @cerner/terra-aggregate-translations package.